clean:documentとhooksのファイル分け #97
